% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_read.R
\name{multi_read}
\alias{multi_read}
\title{Read in multiple data at once and bind them together.}
\usage{
multi_read(
  pattern,
  path = ".",
  bind = FALSE,
  read_func = NULL,
  header = TRUE,
  encoding = "UTF-8",
  ...
)
}
\arguments{
\item{pattern}{The pattern of the data file you want to read.}

\item{path}{The path of the data file you want to read.}

\item{bind}{TRUE if you want to bind the resulting list object into one dataframe.}

\item{read_func}{Type of function you are using to read files.}

\item{header}{TRUE if the header is already in the data.}

\item{encoding}{The way file was encoded.}
}
\description{
Read in multiple data at once and bind them together.
}
